refactor: rename all redux hooks
