+++ src/app/src/components/UserName.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text } from 'react-native';
 import { TMemo } from '@shared/components/TMemo';
-import { useUserName } from '@redux/hooks/useUser';
+import { useUserName } from '@redux/hooks/user';
 
 interface Props {
   uuid: string;
