+++ src/web/components/messageTypes/addons/MsgQuote.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ReplyMsgType } from '@shared/utils/msg-helper';
-import { useUserName } from '@redux/hooks/useUser';
+import { useUserName } from '@redux/hooks/user';
 import { TMemo } from '@shared/components/TMemo';
 import styled from 'styled-components';
 import { UserName } from '@web/components/UserName';
